Extract AbstractPdfIntegrationMockMvcTest class
[#178367884]
